chore: refactor script module and build context system
The old system wasn't maintainable and didn't sit well with me.
Now that the build system is being used in more places, it
really needed a good polish.

This is especially the case since it's being discussed to
release it as a standalone build system later on, un-tied from
Oro.

Needless to say, this fixes a lot of outstanding deficiencies,
and unblocks a number of other features that would have been
a bit painful to implement without this first step.

Closes #24
Closes #23
Closes #22
Closes #18
Closes #17
Closes #12
Closes #10
Closes #2